New module kas-core::event::cx::nav
